relocate service package
