+++ unit-test-coverage.yaml
@@ -38,6 +38,7 @@ packages:
   internal/service/watcher/chartreader: 100
   internal/service/watcher/gateway: 94
   internal/service/watcher/resources: 82
+  internal/service/skrclient:  35
 
   internal/util/collections: 87
   pkg/module/sync: 10
